modularize more routes
