+++ src/routes/fetchCompetitionSchedule.ts
@@ -1,0 +1,33 @@
+import UserReturnData from './UserReturnData'
+/**
+ * GET route '/api/fetchCompetitionSchedule'
+ * Allows the application to get all the matches for a given competition.
+ * @param competition is the Competition id: e.g. '2020ilch'.
+ * @returns back to the client let resobj (competition and ) and HTTP Status Code 200 OK.
+ */
+module.exports = (app: any, dbHandler: any) => {
+    app.get('/api/fetchCompetitionSchedule', async (req: any, res:any) => {
+        let val: UserReturnData =  new UserReturnData();
+        const competition: string = String(req.query.competition);
+        try {
+          val.data = await dbHandler.fetchCompetitionSchedule(req.db, competition).catch((e) => { console.error(e); val.err_occur = true; return {};});
+        } catch (e) {
+          console.error(e);
+          val.err_occur = true;
+        }
+        let resobj = null;
+        if (val.err_occur === false) {
+          resobj = {
+            success: true,
+            competition,
+            data: val.data.data,
+          };
+        } else {
+          resobj = {
+            success: false,
+            reasons: val.err_reasons,
+          };
+        }
+        res.json(resobj);
+    });
+}
